+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [0.24.0] - 2023-09-26
+
 ### Added
 - Command for publishing assets.
 - Show alert if assets are not up-to-date